Fix for single-feature regression.
After the latest set of merges, I found that a single feature
was not rendering on the map. The check for "source.data.features"
was not properly testing for the case where there could be a single
feature stored in "data".  I've updated teh check and it appears to
be working fine now.