More cleaning up of the ref/grouping.
Discovered a bug where the "diffs" for the update cycle
were not happening properly.  This was due to "type: ref" layers
being hydrated out of cycle.

While the "applyStyle" function can handle ref layers properly, the
layer that is referenced can be in a separate OL layer due to grouping.
When that happens applyStyle will not have the layers it needs to
resolve the "ref".

This refactors how layers were being hydrated from their refs.